+++ pkg/tagger/collectors/kubernetes_metadata_mapper.go
@@ -10,10 +10,6 @@ package collectors
 import (
 	"strings"
 
-	"k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/tagger/utils"
 	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/apiserver"
@@ -108,7 +104,7 @@ func (c *KubeMetadataCollector) addToCacheMetadataMapping(kubeletPodList []*kube
 		return nil
 	}
 
-	podList := &v1.PodList{}
+	reachablePods := make([]*kubelet.Pod, 0)
 	nodeName := ""
 	for _, p := range kubeletPodList {
 		if p.Status.PodIP == "" {
@@ -117,18 +113,7 @@ func (c *KubeMetadataCollector) addToCacheMetadataMapping(kubeletPodList []*kube
 		if nodeName == "" && p.Spec.NodeName != "" {
 			nodeName = p.Spec.NodeName
 		}
-
-		pod := &v1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      p.Metadata.Name,
-				Namespace: p.Metadata.Namespace,
-				UID:       types.UID(p.Metadata.UID),
-			},
-			Status: v1.PodStatus{
-				PodIP: p.Status.PodIP,
-			},
-		}
-		podList.Items = append(podList.Items, *pod)
+		reachablePods = append(reachablePods, p)
 	}
-	return c.apiClient.NodeMetadataMapping(nodeName, podList)
+	return c.apiClient.NodeMetadataMapping(nodeName, reachablePods)
 }
