+++ pkg/util/kubernetes/apiserver/apiserver.go
@@ -146,8 +146,6 @@ func (c *APIClient) connect() error {
 }
 
 // MetadataMapperBundle maps pod names to associated metadata.
-//
-// It is updated by mapServices in services.go.
 type MetadataMapperBundle struct {
 	Services ServicesMapper `json:"services,omitempty"`
 	mapOnIP  bool           // temporary opt-out of the new mapping logic
@@ -158,68 +156,6 @@ func newMetadataMapperBundle() *MetadataMapperBundle {
 	return &MetadataMapperBundle{
 		Services: make(ServicesMapper),
 		mapOnIP:  config.Datadog.GetBool("kubernetes_map_services_on_ip"),
-	}
-}
-
-// NodeMetadataMapping only fetch the endpoints from Kubernetes apiserver and add the metadataMapper of the
-// node to the cache
-// Only called when the node agent computes the metadata mapper locally and does not rely on the DCA.
-func (c *APIClient) NodeMetadataMapping(nodeName string, podList *v1.PodList) error {
-	endpointList, err := c.Cl.CoreV1().Endpoints("").List(metav1.ListOptions{TimeoutSeconds: &c.timeoutSeconds})
-	if err != nil {
-		log.Errorf("Could not collect endpoints from the API Server: %q", err.Error())
-		return err
-	}
-	if endpointList.Items == nil {
-		log.Debug("No endpoints collected from the API server")
-		return nil
-	}
-	log.Debugf("Successfully collected endpoints")
-
-	var node v1.Node
-	var nodeList v1.NodeList
-	node.Name = nodeName
-
-	nodeList.Items = append(nodeList.Items, node)
-
-	processKubeServices(&nodeList, podList, endpointList)
-	return nil
-}
-
-// processKubeServices adds services to the metadataMapper cache, pointer parameters must be non nil
-func processKubeServices(nodeList *v1.NodeList, podList *v1.PodList, endpointList *v1.EndpointsList) {
-	if nodeList.Items == nil || podList.Items == nil || endpointList.Items == nil {
-		return
-	}
-	log.Debugf("Identified: %d node, %d pod, %d endpoints", len(nodeList.Items), len(podList.Items), len(endpointList.Items))
-	for _, node := range nodeList.Items {
-		nodeName := node.Name
-		nodeNameCacheKey := cache.BuildAgentKey(metadataMapperCachePrefix, nodeName)
-		freshness := cache.BuildAgentKey(metadataMapperCachePrefix, nodeName, "freshness")
-
-		metaBundle, found := cache.Cache.Get(nodeNameCacheKey)       // We get the old one with the dead pods. if diff reset metabundle and deleted key. Then compute again.
-		freshnessCache, freshnessFound := cache.Cache.Get(freshness) // if expired, freshness not found deal with that
-
-		if !found {
-			metaBundle = newMetadataMapperBundle()
-			cache.Cache.Set(freshness, len(podList.Items), metadataMapExpire)
-		}
-
-		// We want to churn the cache every `metadataMapExpire` and if the number of entries varies between 2 runs..
-		// If a pod is killed and rescheduled during a run, we will only keep the old entry for another run, which is acceptable.
-		if found && freshnessCache != len(podList.Items) || !freshnessFound {
-			cache.Cache.Delete(nodeNameCacheKey)
-			metaBundle = newMetadataMapperBundle()
-			cache.Cache.Set(freshness, len(podList.Items), metadataMapExpire)
-			log.Debugf("Refreshing cache for %s", nodeNameCacheKey)
-		}
-
-		err := metaBundle.(*MetadataMapperBundle).mapServices(nodeName, *podList, *endpointList)
-		if err != nil {
-			log.Errorf("Could not map the services on node %s: %s", node.Name, err.Error())
-			continue
-		}
-		cache.Cache.Set(nodeNameCacheKey, metaBundle, metadataMapExpire)
 	}
 }
